Limit header content's width to 1200px
This width can be any appropriate number, 1200px is an estimate. The
provided designs don't specify an explicit value but content throughout
the full design image is visibly limited within a maximum width. I have
used an extra div container with `.content` class for this purpose.
Thus similar containers can be used for each section with the same class
to wrap the content at a maximum width.